requirements.txt
requests
python-telegram-bot
aiohttp
parsimonious
eth-abi
eth-account